+++ index.js
@@ -13,12 +13,30 @@ module.exports = function PostGraphileConnectionFilterPlugin(
   configOptions
 ) {
   builder.hook("build", build => {
-    // Register plugin version on build
     const pkg = require("./package.json");
+
+    // Check dependencies
     if (!build.versions) {
-      build.versions = {};
+      throw new Error(
+        `Plugin ${pkg.name}@${pkg.version} requires graphile-build@^4.1.0 in order to check dependencies (current version: ${build.graphileBuildVersion})`
+      );
     }
+    const depends = (name, range) => {
+      if (!build.hasVersion(name, range)) {
+        throw new Error(
+          `Plugin ${pkg.name}@${pkg.version} requires ${name}@${range} (${
+            build.versions[name]
+              ? `current version: ${build.versions[name]}`
+              : "not found"
+          })`
+        );
+    }
+    };
+    depends("graphile-build-pg", "^4.5.0");
+
+    // Register this plugin
     build.versions = build.extend(build.versions, { [pkg.name]: pkg.version });
+
     return build;
   });